# Timbangan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timbangan Digital IoT API",
        "description": "API untuk integrasi perangkat timbangan IoT (Arduino/ESP8266) dengan sistem penimbangan digital. Semua endpoint menggunakan token perangkat untuk autentikasi.",
        "contact": {
            "name": "Timbangan Digital Team"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": " Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/incoming-rmpm/settings": {
            "get": {
                "tags": [
                    "Incoming RMPM"
                ],
                "summary": "Get pengaturan sesi aktif timbangan Incoming RMPM",
                "description": "Mendapatkan informasi sesi aktif untuk penimbangan bahan baku RMPM masuk.",
                "operationId": "b13cd4442634dbfe18ee7f6b938d29eb",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "DEV-TOKEN-RMPM-001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ready",
                                                "idle"
                                            ],
                                            "example": "ready"
                                        },
                                        "kode_produksi": {
                                            "type": "string",
                                            "example": "BATCH-001"
                                        },
                                        "nama_produk": {
                                            "type": "string",
                                            "example": "Tepung Tapioka"
                                        },
                                        "operator": {
                                            "type": "string",
                                            "example": "Budi"
                                        },
                                        "no_surat": {
                                            "type": "string",
                                            "example": "SJ-RMPM-001"
                                        },
                                        "nama_supplier": {
                                            "type": "string",
                                            "example": "PT. Supplier RMPM"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/incoming-rmpm/weight": {
            "post": {
                "tags": [
                    "Incoming RMPM"
                ],
                "summary": "Kirim data berat timbangan Incoming RMPM",
                "description": "Menyimpan data berat RMPM masuk. Membutuhkan sesi aktif yang sudah diinput oleh operator.",
                "operationId": "baa6f256510ca08cfcad66b7bc707822",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "token",
                                    "weight"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "DEV-TOKEN-RMPM-001"
                                    },
                                    "weight": {
                                        "description": "Berat dalam kg",
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berat berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WeightSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Berat tidak valid / Tidak ada sesi aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "WeightSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Data berat 25.5 kg tersimpan"
                    },
                    "record_id": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            }
        }
    },
...
}